\name{crash_taxonomies}
\docType{data}
\alias{crash_taxonomies}
\title{Taxonomies to illustrate the functionality of meltt}
\description{
These taxonomies formalize how the information about model, color, and type of accident in our three artificial car crash datasets map onto one another.
}
\usage{data(crashMD)}
\format{A list of three \code{data.frame} containing information about the different categories (specific to general) of models, colors and degree of damages coded in each dataset.}
\source{Simulated data.}
\references{
Karsten Donnay, Eric T. Dunford, Erin C. McGrath, David Backer, David E. Cunningham. (2018). "Integrating Conflict Event Data." \emph{Journal of Conflict Resolution}.
}
\author{
  Karsten Donnay and Eric Dunford.
}
