\name{print.meltt}
\alias{print.meltt}
\title{
Print function for objects of class meltt. 
}
\description{
Overloads the default \code{print()} for objects of class \code{meltt}.
}
\usage{
\method{print}{meltt}(x, ...)
}
\arguments{
  \item{x}{
  object of class \code{\link{meltt}}.}
  \item{...}{
   further arguments passed to or from other methods.}
}
\references{
Karsten Donnay, Eric Dunford, Erin C. McGrath, David Backer, David E. Cunningham. (2016). "MELTT: Matching Event Data by Location, Time and Type." Paper presented at \emph{Annual Conference of the Midwest Political Science Association}, April 7-10 2016, Chicago.
}
\author{
  Karsten Donnay and Eric Dunford.
}
\seealso{
  \code{\link{meltt}}
}