\name{crash_data3}
\docType{data}
\alias{crash_data3}
\title{Dataset to illustrate the functionality of meltt}
\description{
  This artificial dataset illustrates how \pkg{meltt} can be used to automatically integrate and disambiguate event data. It contains timing and location information about (simulated) car crashes for one month (Jan. 2012) in the state of Maryland, U.S., with information about the model, color and, and type of accident.
}
\usage{data(crashMD)}
\format{A \code{data.frame} containing observations.}
\source{Simulated data.}
\references{
Karsten Donnay, Eric T. Dunford, Erin C. McGrath, David Backer, David E. Cunningham. (2018). "Integrating Conflict Event Data." \emph{Journal of Conflict Resolution}.
}
\author{
  Karsten Donnay and Eric Dunford.
}
