\name{crash_taxonomies}
\docType{data}
\alias{crash_taxonomies}
\title{Taxonomies to illustrate the functionality of meltt}
\description{
These taxonomies formalize how the information about model, color, and type of accident in our three artificial car crash datasets map onto one another.
}
\usage{data(crashMD)}
\format{A list of three \code{data.frame} containing information about the different categories (specific to general) of models, colors and degree of damages coded in each dataset.}
\source{Simulated data.}
\references{
Karsten Donnay, Eric Dunford, Erin C. McGrath, David Backer, David E. Cunningham. (2016). "MELTT: Matching Event Data by Location, Time and Type." Paper presented at \emph{Annual Conference of the Midwest Political Science Association}, April 7-10 2016, Chicago.
}
\author{
  Karsten Donnay and Eric Dunford.
}
