\name{print.meltt}
\alias{print.meltt}
\title{
Print function for objects of class meltt. 
}
\description{
Overloads the default \code{print()} for objects of class \code{meltt}.
}
\usage{
\method{print}{meltt}(x, ...)
}
\arguments{
  \item{x}{
  object of class \code{\link{meltt}}.}
  \item{...}{
   further arguments passed to or from other methods.}
}
\references{
Karsten Donnay, Eric T. Dunford, Erin C. McGrath, David Backer, David E. Cunningham. (2018). "Integrating Conflict Event Data." \emph{Journal of Conflict Resolution}.
}
\author{
  Karsten Donnay and Eric Dunford.
}
\seealso{
  \code{\link{meltt}}
}